Refactor how we set the ipv6 proxy
This patch has been contributed by a community member
