+++ app/src/main/java/org/openobservatory/ooniprobe/common/ProxySettings.java
@@ -1,7 +1,12 @@
 package org.openobservatory.ooniprobe.common;
 
+import com.google.common.net.InetAddresses;
+
+import java.net.Inet6Address;
+import java.net.InetAddress;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.net.UnknownHostException;
 
 /**
  * ProxySettings contains the settings configured inside the proxy. Please, see the
@@ -66,6 +71,22 @@ public class ProxySettings {
         pm.setProxyPort(port);
     }
 
+    // isIPv6 tells us whether the input is an IPv6 address.
+    @SuppressWarnings("UnstableApiUsage")
+    private boolean isIPv6(String hostname) {
+        try {
+            // The UnknownHostException may suggest we're doing a domain
+            // name resolution here. While this is possible in theory, we
+            // check whether we're dealing with a literal IP address in
+            // advance. In such a case getByName won't attempt to resolve
+            // anything because it's already given an IP address.
+            return InetAddresses.isInetAddress(hostname) &&
+                    InetAddress.getByName(hostname) instanceof Inet6Address;
+        } catch (UnknownHostException e) {
+            return false;
+        }
+    }
+
     /** getProxyString returns to you the proxy string you should pass to oonimkall. */
     public String getProxyString() throws URISyntaxException {
         if (protocol == Protocol.NONE)
@@ -76,6 +97,9 @@ public class ProxySettings {
             // Alright, we now need to construct a new SOCKS5 URL. We are going to defer
             // doing that to the Java standard library (er, the Android stdlib).
             String urlStr = "socks5://" + hostname + ":" + port + "/";
+            if (isIPv6(hostname)) {
+                urlStr = "socks5://[" + hostname + "]:" + port + "/"; // IPv6 must be quoted in URLs
+            }
             URI url = new URI(urlStr);
             return url.toASCIIString();
         }
@@ -84,10 +108,6 @@ public class ProxySettings {
 
     /** InvalidProxyURL indicates that the proxy URL is not valid. */
     public static class InvalidProxyURL extends Exception {
-        InvalidProxyURL(String message, Throwable cause) {
-            super(message, cause);
-        }
-
         InvalidProxyURL(String message) {
             super(message);
         }
